refactor: create reusable quiz table and list components
- Extract QuizTableRow component from monolithic table in quizzes page
- Create QuizTable component to wrap table structure and header
- Add QuizListCard component for flexible quiz card display with customizable action buttons
- Refactor QuizReviewPanel to use new QuizListCard component
- Reduce quizzes.tsx from 217 lines to 92 lines (58% reduction)
- Reduce QuizReviewPanel from 193 lines to 79 lines (59% reduction)
- Add reusable components to common exports for consistent quiz data display
- All components are memoized for performance optimization
- Fix router navigation to properly handle dynamic route parameters
- Remove invalid search parameters that don't exist in quiz detail route
- Maintain consistent styling and functionality across all quiz displays
