+++ frontend/src/components/Common/QuizTableRow.tsx
@@ -1,0 +1,77 @@
+import { Badge, HStack, Table, Text, VStack } from "@chakra-ui/react"
+import { Link as RouterLink } from "@tanstack/react-router"
+import { memo } from "react"
+
+import type { Quiz } from "@/client/types.gen"
+import { Button } from "@/components/ui/button"
+import { StatusLight } from "@/components/ui/status-light"
+import { formatDate, getQuizStatusText, getSelectedModulesCount } from "@/lib/utils"
+import { PROCESSING_STATUSES } from "@/lib/constants"
+
+interface QuizTableRowProps {
+  quiz: Quiz
+}
+
+export const QuizTableRow = memo(function QuizTableRow({
+  quiz,
+}: QuizTableRowProps) {
+  const moduleCount = getSelectedModulesCount(quiz)
+
+  return (
+    <Table.Row key={quiz.id}>
+      <Table.Cell>
+        <VStack align="start" gap={1}>
+          <Text fontWeight="medium">{quiz.title}</Text>
+          <Text fontSize="sm" color="gray.500">
+            {moduleCount} module{moduleCount !== 1 ? "s" : ""} selected
+          </Text>
+        </VStack>
+      </Table.Cell>
+      <Table.Cell>
+        <VStack align="start" gap={1}>
+          <Text>{quiz.canvas_course_name}</Text>
+          <Text fontSize="sm" color="gray.500">
+            ID: {quiz.canvas_course_id}
+          </Text>
+        </VStack>
+      </Table.Cell>
+      <Table.Cell>
+        <Badge variant="solid" colorScheme="blue">
+          {quiz.question_count}
+        </Badge>
+      </Table.Cell>
+      <Table.Cell>
+        <Badge variant="outline" colorScheme="purple">
+          {quiz.llm_model}
+        </Badge>
+      </Table.Cell>
+      <Table.Cell>
+        <HStack gap={2} align="center">
+          <StatusLight
+            extractionStatus={
+              quiz.content_extraction_status || PROCESSING_STATUSES.PENDING
+            }
+            generationStatus={
+              quiz.llm_generation_status || PROCESSING_STATUSES.PENDING
+            }
+          />
+          <Text fontSize="sm" color="gray.600">
+            {getQuizStatusText(quiz)}
+          </Text>
+        </HStack>
+      </Table.Cell>
+      <Table.Cell>
+        <Text fontSize="sm">
+          {quiz.created_at ? formatDate(quiz.created_at) : "Unknown"}
+        </Text>
+      </Table.Cell>
+      <Table.Cell>
+        <HStack gap={2}>
+          <Button size="sm" variant="outline" asChild>
+            <RouterLink to={`/quiz/${quiz.id}`}>View</RouterLink>
+          </Button>
+        </HStack>
+      </Table.Cell>
+    </Table.Row>
+  )
+})
