+++ frontend/src/components/Common/index.ts
@@ -3,3 +3,8 @@ export * from './LoadingSkeleton'
 export * from './EmptyState'
 export * from './ErrorState'
 export * from './PageHeader'
+
+// Quiz display components
+export * from './QuizTable'
+export * from './QuizTableRow'
+export * from './QuizListCard'
