+++ frontend/src/components/dashboard/panels/QuizReviewPanel.tsx
@@ -10,10 +10,9 @@ import { Link as RouterLink } from "@tanstack/react-router"
 
 import type { Quiz } from "@/client/types.gen"
 import { Button } from "@/components/ui/button"
-import { StatusLight } from "@/components/ui/status-light"
-import { EmptyState, LoadingSkeleton } from "@/components/common"
+import { EmptyState, LoadingSkeleton, QuizListCard } from "@/components/common"
 import { getQuizzesNeedingReview } from "@/lib/utils"
-import { UI_SIZES, UI_TEXT, PROCESSING_STATUSES } from "@/lib/constants"
+import { UI_SIZES, UI_TEXT } from "@/lib/constants"
 
 interface QuizReviewPanelProps {
   quizzes: Quiz[]
@@ -51,80 +50,16 @@ export function QuizReviewPanel({ quizzes, isLoading }: QuizReviewPanelProps) {
         ) : (
           <VStack gap={4} align="stretch">
             {reviewQuizzes.slice(0, UI_SIZES.PANEL.MAX_ITEMS + 1).map((quiz) => (
-              <Box
+              <QuizListCard
                 key={quiz.id}
-                p={3}
-                border="1px solid"
-                borderColor="gray.200"
-                borderRadius="md"
-                bg="gray.50"
-                _hover={{ bg: "gray.100" }}
-                transition="background-color 0.2s"
-              >
-                <VStack align="stretch" gap={2}>
-                  <VStack align="start" gap={1}>
-                    <HStack justify="space-between" align="start" width="100%">
-                      <Text
-                        fontWeight="medium"
-                        fontSize="sm"
-                        lineHeight="tight"
-                        truncate
-                      >
-                        {quiz.title}
-                      </Text>
-                      <StatusLight
-                        extractionStatus={
-                          quiz.content_extraction_status || PROCESSING_STATUSES.PENDING
-                        }
-                        generationStatus={
-                          quiz.llm_generation_status || PROCESSING_STATUSES.PENDING
-                        }
-                      />
-                    </HStack>
-                    <Text fontSize="xs" color="gray.600" truncate>
-                      {quiz.canvas_course_name}
-                    </Text>
-                  </VStack>
-
-                  <HStack
-                    justify="space-between"
-                    align="center"
-                    gap={2}
-                    flexWrap="wrap"
-                  >
-                    <HStack gap={2} flex="1" minW="0">
-                      <Badge
-                        variant="solid"
-                        colorScheme="blue"
-                        size="sm"
-                        flexShrink={0}
-                      >
-                        {quiz.question_count} questions
-                      </Badge>
-                      {quiz.llm_model && (
-                        <Badge
-                          variant="outline"
-                          colorScheme="purple"
-                          size="sm"
-                          flexShrink={0}
-                        >
-                          {quiz.llm_model}
-                        </Badge>
-                      )}
-                    </HStack>
-
-                    <Button size="sm" variant="outline" asChild flexShrink={0}>
-                      <RouterLink
-                        to="/quiz/$id"
-                        params={{ id: quiz.id || "" }}
-                        search={{ tab: "questions" }}
-                      >
-                        Review
-                      </RouterLink>
-                    </Button>
-                  </HStack>
-                </VStack>
-              </Box>
+                quiz={quiz}
+                actionButton={{
+                  text: "Review",
+                  to: "/quiz/$id",
+                  params: { id: quiz.id || "" }
+                }}
+                compact
+              />
             ))}
 
             {reviewQuizzes.length > 5 && (
